# Reduce memory usage in static graph build by snapshotting the project instance