… src/Build.UnitTests/Graph/ProjectGraph_Tests.cs
… src/Build.UnitTests/ProjectCache/ProjectCacheTests.cs
… src/Build/Graph/GraphBuildSubmission.cs
… src/Build/Graph/GraphBuilder.cs
… src/Build/Graph/ProjectGraph.cs
… src/Build/Graph/ProjectGraphNode.cs
… src/Build/Graph/ProjectInterpretation.cs