### Context
When building with Static Graph, it evaluates the entire solution space to computes the most optimal path for parallelization.  However, it keeps those evaluated instances in memory in hope they might be reused.  From my testing, these are not reused because new instances are created with different properties from platform negotiation. 

### Changes Made
Create a snapshot of the data that Static Graph would use then "free" those instances.

### Testing
Add after printing `Static graph loaded` message.
```
GC.Collect();
LogMessage($"GraphBuild GC: {GC.GetTotalMemory(false)} bytes");
```
**Solution 1:**
*baseline:*
Static graph loaded in 37.884 seconds: 166 nodes, 1628 edges
GraphBuild GC: **185,411,136** bytes

*new:*
Static graph loaded in 37.024 seconds: 166 nodes, 1628 edges
GraphBuild GC: **24,743,840** bytes


**Solution 2:**
*baseline*
Static graph loaded in 498.914 seconds: 913 nodes, 12862 edges
GraphBuild GC: **1,141,466,552** bytes 

*new*
Static graph loaded in 425.508 seconds: 913 nodes, 12862 edges
GraphBuild GC: **62,436,160** bytes 

### Notes
I was hoping there would be more performance uplift from reducing GC pressure.  Static graph performance would need further investigation.
The Snapshot class is mimicking the ProjectInstance API to reduce code churn.  There might be further data structure optimization without that limitation.