@rokonec 2023-08-29
@yuehuang010  I looked at the code and measured it a bit. 
I think saving memory usage, in context on usual build agents used for graph builds, is not as useful as saving allocations.

The premise that evaluation during graph build is not reused is not valid when project is not multitargeted. In my test project it caused redundant re-evaluations. Since we plan to eventually use and support graph build for external customers, this would harm everyone non multitargeting.

I have discussed it with @rainersigwald and we believe that graph build shall reuse evaluations from graph building phase, otherwise we risk that actual shape of graph during build differs from shape of graph by which we build.

I think this approach need some further design.